Added new files coordinates.hpp/.cpp, modification of island.hpp - move coordinates class to new file, delete references, modification of operator==
