+++ shm/island.hpp
@@ -1,30 +1,11 @@
 #pragma once
+#include "coordinates.hpp"
 
 class Island {    
 public:
-    Island(int positionX, int positionY)
-    : position_(Coordinates(positionX, positionY))
+    Island(Coordinates position)
+    : position_(position)
     {}
-
-    class Coordinates {
-    public:    
-        Coordinates(int& positionX, int& positionY) 
-        : positionX_(positionX)
-        , positionY_(positionY)
-        {}
-
-        void setPositionX(int& );
-        int getPositionX() const;
-
-        void setPositionY(int& );
-        int getPositionY() const;
-    
-        bool operator== (const Coordinates& n_pos) const;
-
-    private:
-        int positionX_;
-        int positionY_;
-    };
 
     Coordinates getPosition() const;
